Renames component
