+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationCase/HearingArrangements/HearingArrangements.tsx
@@ -44,7 +44,7 @@ import * as constants from '@island.is/judicial-system/consts'
 
 import RequestCourtDate from '../../SharedComponents/RequestCourtDate/RequestCourtDate'
 import SelectCourt from '../../SharedComponents/SelectCourt/SelectCourt'
-import SelectProsecutor from '../../SharedComponents/SelectProsecutor/SelectProsecutor'
+import ProsecutorSection from '../../SharedComponents/ProsecutorSection/ProsecutorSection'
 
 const HearingArrangements = () => {
   const router = useRouter()
@@ -191,7 +191,7 @@ const HearingArrangements = () => {
             <Box component="section" marginBottom={5}>
               <BlueBox>
                 <Box marginBottom={2}>
-                  <SelectProsecutor onChange={handleProsecutorChange} />
+                  <ProsecutorSection onChange={handleProsecutorChange} />
                 </Box>
                 <Checkbox
                   name="isHeightenedSecurityLevel"
